+++ sql/item_vers.cc
@@ -63,7 +63,7 @@ Item_func_vtq_ts::get_date(MYSQL_TIME *res, ulonglong fuzzy_date)
   null_value= !trt.query(trx_id);
   if (null_value)
   {
-    my_error(ER_VERS_NO_TRX_ID, MYF(0), trx_id);
+    my_error(ER_VERS_NO_TRX_ID, MYF(0), (longlong) trx_id);
     return true;
   }
